requirements.txt
faiss-cpu
sentence-transformers
langchain
openai
streamlit
transformers
accelerate
pymupdf
python-docx
tqdm
